Enable the 'Add' button to the people licence page on self-serve
by making the adapter 'editable' (if the org is not exceptional - i.e.
not partnership or sole trader).

Also disable the undelying add/edit/delete actions, but still allow the
edit action for GET request (which was already accessible and more of a
'view' action really).

Delete action will probably be re-enabled later but there is an open
question about whether to allow this directly or only via a director
change variation.